+++ test/data/NetworkPreparation/link.csv
@@ -1,5 +1,6 @@
-a_node,b_node,direction,distance,modes,link_type,name,geometry,link_id
-11,12,1,1.0,a,test,first link,"LINESTRING (0 0, 1 0)",1
-12,13,1,1.0,r,test,second link,"LINESTRING (1 0, 1 1)",2
-13,14,0,1.0,x,test,third link,"LINESTRING (1 1, 0 1)",3
-14,11,-1,1.0,xrc,test,fourth link,"LINESTRING (0 1, 0 0)",4
+a_node,b_node,direction,distance,modes,link_type,name,geometry,link_id,matrix_ab,matrix_ba,matrix_tot
+11,12,1,1.0,a,test,first link,"LINESTRING (0 0, 1 0)",1,50,50,100
+12,13,1,1.0,r,test,second link,"LINESTRING (1 0, 1 1)",2,42,63,105
+13,14,0,1.0,x,test,third link,"LINESTRING (1 1, 0 1)",3,17,18,35
+14,11,-1,1.0,xrc,test,fourth link,"LINESTRING (0 1, 0 0)",4,32,32,64
+11,13,-1,1.2,a,highway,fifth link,"LINESTRING (0 0, 1 1)",5,19,11,30
